create component to load all spots
